requirements.txt
telethon>=1.28.0
aiohttp>=3.8.0
python-dotenv>=1.0.0